create Customer all tests

+++ src/test/java/customer/CreateCustomerTests.java
@@ -17,11 +17,11 @@ import utils.Generator;
 import java.util.Map;
 
 import static org.junit.Assert.*;
+import static utils.Generator.*;
 
 public class CreateCustomerTests {
 
     private CustomerClient customerClient = new CustomerClient();
-    Response response;
     private String token;
     @Before
     public  void setUp(){
@@ -36,10 +36,7 @@ public class CreateCustomerTests {
 
         assertEquals("Неверный статус код", HttpStatus.SC_OK, response.statusCode());
 
-        String responseBody = response.getBody().asString();
-        CustomerToken authResponse = CustomerToken.fromJson(responseBody);
-        token = authResponse.getAccessToken();
-
+        token = CustomerToken.extractAccessToken(response);
     }
     @Test
     @DisplayName("Check response body structure after authorization")
@@ -53,15 +50,12 @@ public class CreateCustomerTests {
         assertNotNull(jsonPath.getString("accessToken"));
         assertNotNull(jsonPath.getString("refreshToken"));
 
-        // Проверяем структуру объекта "user"
         Map<String, String> user = jsonPath.getMap("user");
         assertNotNull(user);
         assertTrue(user.containsKey("email"));
         assertTrue(user.containsKey("name"));
 
-        String responseBody = response.getBody().asString();
-        CustomerToken authResponse = CustomerToken.fromJson(responseBody);
-        token = authResponse.getAccessToken();
+        token = CustomerToken.extractAccessToken(response);
     }
     @Test
     @DisplayName("Check status code create duplicate customer")
@@ -69,18 +63,75 @@ public class CreateCustomerTests {
         Customer customer = Generator.randomCustomer();
         Response response = customerClient.create(customer);
 
-        String responseBody = response.getBody().asString();
-        CustomerToken authResponse = CustomerToken.fromJson(responseBody);
-        token = authResponse.getAccessToken();
+        token = CustomerToken.extractAccessToken(response);
 
         Response responseDuplicate = customerClient.create(customer);
 
-        assertEquals("Создание дубликата курьера", HttpStatus.SC_FORBIDDEN, responseDuplicate.statusCode());
+        assertEquals("Создание дубликата клиента", HttpStatus.SC_FORBIDDEN, responseDuplicate.statusCode());
+    }
+    @Test
+    @DisplayName("Check response body after try duplicate authorization")
+    public void testCreateCustomerDuplicateWithResponse() {
 
+        Customer customer = Generator.randomCustomer();
+        Response response = customerClient.create(customer);
+        token = CustomerToken.extractAccessToken(response);
+        Response responseDuplicate = customerClient.create(customer);
+
+        JsonPath jsonPath = responseDuplicate.jsonPath();
+        assertFalse(jsonPath.getBoolean("success"));
+        String errorMessage = jsonPath.getString("message");
+        assertEquals("Неверное сообщение об ошибке", "User already exists", errorMessage);
+    }
+    @Test
+    @DisplayName("Check status code and body after authorization with out name")
+    public void testCreateCustomerWithOutName() {
+        Customer customer = new Customer()
+                .withEmail(randomEmail())
+                .withPassword(randomPassword());
+        Response response = customerClient.create(customer);
+
+        assertEquals("Неверный статус код", HttpStatus.SC_FORBIDDEN, response.statusCode());
+
+        JsonPath jsonPath = response.jsonPath();
+        assertFalse(jsonPath.getBoolean("success"));
+        String errorMessage = jsonPath.getString("message");
+        assertEquals("Неверное сообщение об ошибке", "Email, password and name are required fields", errorMessage);
+    }
+    @Test
+    @DisplayName("Check status code and body after authorization with out email")
+    public void testCreateCustomerWithOutEmail() {
+        Customer customer = new Customer()
+                .withEmail(randomName())
+                .withPassword(randomPassword());
+        Response response = customerClient.create(customer);
+
+        assertEquals("Неверный статус код", HttpStatus.SC_FORBIDDEN, response.statusCode());
+
+        JsonPath jsonPath = response.jsonPath();
+        assertFalse(jsonPath.getBoolean("success"));
+        String errorMessage = jsonPath.getString("message");
+        assertEquals("Неверное сообщение об ошибке", "Email, password and name are required fields", errorMessage);
+    }
+    @Test
+    @DisplayName("Check status code and body after authorization with out password")
+    public void testCreateCustomerWithOutPassword() {
+        Customer customer = new Customer()
+                .withPassword(randomName())
+                .withEmail(randomEmail());
+        Response response = customerClient.create(customer);
+
+        assertEquals("Неверный статус код", HttpStatus.SC_FORBIDDEN, response.statusCode());
+
+        JsonPath jsonPath = response.jsonPath();
+        assertFalse(jsonPath.getBoolean("success"));
+        String errorMessage = jsonPath.getString("message");
+        assertEquals("Неверное сообщение об ошибке", "Email, password and name are required fields", errorMessage);
     }
     @After
     public void deleteCustomer(){
-
+        if (token != null){
         customerClient.delete(token);
+        }
     }
 }
